コンポーネントを分割する
