+++ src/app/_components/PopulationChart/Presenter..tsx
@@ -1,0 +1,36 @@
+import { components } from '@/lib/api/schema'
+import {
+  LineChart,
+  CartesianGrid,
+  XAxis,
+  YAxis,
+  Tooltip,
+  Legend,
+  Line,
+} from 'recharts'
+
+type Props = {
+  dataList: {
+    prefecture: components['schemas']['Prefecture']
+    data: components['schemas']['Population'][]
+  }[]
+}
+export function PopulationChartPresenter({ dataList }: Props) {
+  return (
+    <LineChart width={800} height={300}>
+      <CartesianGrid strokeDasharray='3 3' />
+      <XAxis dataKey='year' type='category' allowDuplicatedCategory={false} />
+      <YAxis dataKey='value' />
+      <Tooltip />
+      <Legend />
+      {dataList.map(({ prefecture, data }) => (
+        <Line
+          key={prefecture.prefCode}
+          dataKey='value'
+          data={data}
+          name={prefecture.prefName}
+        />
+      ))}
+    </LineChart>
+  )
+}
